flask/TechTest/backend/requirements.txt
Flask[async]
flask-sqlalchemy
pydantic-settings
pymysql
cryptography
numpy